Move ITask to /shared/models, #57
